Changed the project to maven style and moved the files to their respective folders
Optimized fxml to remove unnecessary children
